Refactor device wearer integration tests
