+++ integration_tests/pages/components/forms/about-the-device-wearer/deviceWearerForm.ts
@@ -93,8 +93,6 @@ export default class AboutDeviceWearerFormComponent extends FormComponent {
     return new FormInputComponent(this.form, label)
   }
 
-  disabilities
-
   // Interpreter
 
   get interpreterRequiredField(): FormRadiosComponent {
